requirements.txt
clip
pillow
numpy